+++ src/app/admin/pages/admin-create-or-update-product/admin-create-or-update-product.component.spec.ts
@@ -1,0 +1,24 @@
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { IonicModule } from '@ionic/angular';
+
+import { AdminCreateOrUpdateProductComponent } from './admin-create-or-update-product.component';
+
+describe('AdminCreateOrUpdateProductComponent', () => {
+  let component: AdminCreateOrUpdateProductComponent;
+  let fixture: ComponentFixture<AdminCreateOrUpdateProductComponent>;
+
+  beforeEach(waitForAsync(() => {
+    TestBed.configureTestingModule({
+      declarations: [ AdminCreateOrUpdateProductComponent ],
+      imports: [IonicModule.forRoot()]
+    }).compileComponents();
+
+    fixture = TestBed.createComponent(AdminCreateOrUpdateProductComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  }));
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});