+++ src/app/admin/pages/admin-list-products/admin-list-products.component.ts
@@ -1,0 +1,14 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-admin-list-products',
+  templateUrl: './admin-list-products.component.html',
+  styleUrls: ['./admin-list-products.component.scss'],
+})
+export class AdminListProductsComponent  implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {}
+
+}