+++ src/app/client/pages/client-detail-product/client-detail-product.component.spec.ts
@@ -1,0 +1,24 @@
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { IonicModule } from '@ionic/angular';
+
+import { ClientDetailProductComponent } from './client-detail-product.component';
+
+describe('ClientDetailProductComponent', () => {
+  let component: ClientDetailProductComponent;
+  let fixture: ComponentFixture<ClientDetailProductComponent>;
+
+  beforeEach(waitForAsync(() => {
+    TestBed.configureTestingModule({
+      declarations: [ ClientDetailProductComponent ],
+      imports: [IonicModule.forRoot()]
+    }).compileComponents();
+
+    fixture = TestBed.createComponent(ClientDetailProductComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  }));
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});