+++ CHANGELOG.md
@@ -1,7 +1,48 @@
 Changelog
+## [1.4.0](https://github.com/darksoul333666/Straapberry/compare/v1.3.1...v1.4.0) (2024-07-27)
+
+
+### ✨ Features
+
+* modules, routing and pages for authentication were added ([a2a0b5f](https://github.com/darksoul333666/Straapberry/commit/a2a0b5f0df848ddac0444eb7acde0e52ef4a52c3))
+
+### [1.3.1](https://github.com/darksoul333666/Straapberry/compare/v1.3.0...v1.3.1) (2024-07-27)
+
+
+### 🐛 Bug Fixes
+
+* changing repository name ([ee644b5](https://github.com/darksoul333666/Straapberry/commit/ee644b53faa4d506435c38dc01763246664efb88))
+
+## [1.3.0](https://github.com/darksoul333666/Straapberry/compare/v1.2.0...v1.3.0) (2024-07-27)
+
+
+### ✨ Features
+
+* changing carpet name ([2a7eec7](https://github.com/darksoul333666/Straapberry/commit/2a7eec7d6e92ff13a6bfc41b5de7a28999a2f4b9))
+* client modules pages and routing were added ([1a8ef11](https://github.com/darksoul333666/Straapberry/commit/1a8ef11ae2c733ef3e3cee8f3b5bedd748c82254))
+
+## [1.2.0](https://github.com/darksoul333666/Straapberry/compare/v1.1.0...v1.2.0) (2024-07-27)
+
+### ✨ Features
+
+* routing and module for admin added ([7a5962b](https://github.com/darksoul333666/Straapberry/commit/7a5962b971d82ecb954e4b7ba39cda9c3e9f8643))
+
+## 1.1.0 (2024-07-27)
+
+
+### 🚚 Chores
+
+* **release:** 0.0.1 ([f347e8a](https://github.com/darksoul333666/Straapberry/commit/f347e8a56da6c7dc0569e8a40fdd71fe12848c28))
+
+
+### ✨ Features
+
+* configuration to conventional commits and standar versioning ([cad183c](https://github.com/darksoul333666/Straapberry/commit/cad183c6a36048395e442ad3ab0c9990b6bc4c55))
+* generation admin modules ([92b0d67](https://github.com/darksoul333666/Straapberry/commit/92b0d6758d5c106096b1bbb6347f67425d10133d))
+
 ## 1.0.0 (2024-07-27)
 
 
 ### ✨ Features
 
-* configuration to conventional commits and standar versioning ([cad183c](https://github.com/lumston/lumston-medusa-typescript-angular-vristo-frontend/commit/cad183c6a36048395e442ad3ab0c9990b6bc4c55))
+* configuration to conventional commits and standar versioning ([cad183c](https://github.com/darksoul333666/Straapberry/commit/cad183c6a36048395e442ad3ab0c9990b6bc4c55))